Phase 3: Extract modals and helpers from large page/component files
Decompose the 4 largest page/component files by extracting self-contained
modals and pure helper functions into focused, single-responsibility files.
No behavioral changes — purely structural reorganization.

SchedulePage.jsx (2,017 → 1,370 lines, -32%):
- Extract ScoresDetailsModal.jsx (377 lines): fullscreen scores modal with
  all score rendering helpers (rankPointDisplay, scoreAchieved, addScoreType,
  scoresRow, expandScoresRow)
- Extract AdjustAlliancesModal.jsx (177 lines): alliance editing modal with
  internal formData state and save logic
- Extract scheduleParsingHelpers.js (154 lines): pure CSV/XLSX parsing and
  normalization functions

TopButtons.jsx (1,458 → 1,038 lines, -29%):
- Extract AdHocMatchModal.jsx (258 lines): ad-hoc match team selection modal
  with adHocStation helper
- Extract allianceHelpers.js (196 lines): 7 pure alliance helper functions

RanksPage.jsx (1,159 → 989 lines, -15%):
- Extract AdvancementDetailsModal.jsx (198 lines): championship advancement
  details modal replacing 170-line IIFE

TeamDataPage.jsx (1,342 → 1,074 lines, -20%):
- Extract TeamEditModal.jsx (264 lines): team data editing form with 13
  local field states
- Extract TeamHistoryModal.jsx (86 lines): pure presentational history modal

Design: open/close state stays in parent, modals receive onSave/onRestore
callbacks not raw setters, score helpers kept with their modal component.

12 files changed, 8 new files, ~1500 lines redistributed. Build: zero warnings.
