+++ src/pages/RanksPage.jsx
@@ -1,14 +1,13 @@
 import {
   Alert,
-  Button,
   Container,
-  Modal,
   Table,
   Row,
   Col,
   OverlayTrigger,
   Tooltip,
 } from "react-bootstrap";
+import AdvancementDetailsModal, { getRegionalAdvancementTotal } from "components/AdvancementDetailsModal";
 import {
   SortNumericDown,
   SortNumericUp,
@@ -217,11 +216,6 @@ function RanksPage({
     return team?.nameShortLocal ? team?.nameShortLocal : team?.nameShort;
   }
 
-  /** Per-team API sometimes returns totalPoints: 0; derive total from regional point fields when needed. */
-  function getRegionalAdvancementTotal(r) {
-    if (r.totalPoints != null && r.totalPoints > 0) return r.totalPoints;
-    return (r.regional1Points ?? 0) + (r.regional2Points ?? 0) + (r.regionalDirectPoints ?? 0);
-  }
 
   const isLeagueMeet = ftcMode && selectedEvent?.value?.type === "1";
 
@@ -981,177 +975,13 @@ function RanksPage({
           </div>
         )}
 
-      {(() => {
-        const d = advancementDetailTeam?.districtRankDetails;
-        const r = advancementDetailTeam?.regionalAdvancement;
-        let headerClass = "bg-light text-dark";
-        let closeVariant = undefined;
-        if (advancementDetailType === "district") {
-          if (d?.qualifiedFirstCmp) {
-            headerClass = "bg-success text-white";
-            closeVariant = "white";
-          } else if (d?.qualifiedDistrictCmp) {
-            headerClass = "bg-info text-white";
-            closeVariant = "white";
-          } else {
-            headerClass = "bg-warning text-dark";
-          }
-        } else {
-          const qualified = r?.qualifiedFirstCmp;
-          const hasNotPlayedBoth = r && !r.regional2Details;
-          if (qualified) {
-            headerClass = "bg-success text-white";
-            closeVariant = "white";
-          } else if (hasNotPlayedBoth) {
-            headerClass = "bg-warning text-dark";
-          }
-        }
-        return (
-      <Modal
-        centered
+      <AdvancementDetailsModal
         show={showAdvancementModal}
         onHide={handleCloseAdvancementModal}
-        size="lg"
-      >
-        <Modal.Header closeButton className={headerClass} closeVariant={closeVariant}>
-          <Modal.Title>
-            Championship Advancement — Team{" "}
-            {remapNumberToString
-              ? remapNumberToString(advancementDetailTeam?.teamNumber)
-              : advancementDetailTeam?.teamNumber}{" "}
-            ({advancementDetailType === "district" ? "District" : "World"})
-          </Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {advancementDetailType === "district" &&
-            advancementDetailTeam?.districtRankDetails && (
-              <Container fluid>
-                <p>
-                  <b>District Rank:</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.rank} (
-                  {advancementDetailTeam.districtRankDetails.totalPoints} pts)
-                </p>
-                <p>
-                  <b>Age points:</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.teamAgePoints ?? 0}
-                </p>
-                <p>
-                  <b>Event 1 ({advancementDetailTeam.districtRankDetails.event1Code}):</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.event1Points ?? 0} pts
-                </p>
-                <p>
-                  <b>Event 2 ({advancementDetailTeam.districtRankDetails.event2Code}):</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.event2Points ?? 0} pts
-                </p>
-                {advancementDetailTeam.districtRankDetails.districtCmpCode && (
-                  <p>
-                    <b>DCMP ({advancementDetailTeam.districtRankDetails.districtCmpCode}):</b>{" "}
-                    {advancementDetailTeam.districtRankDetails.districtCmpPoints ?? 0} pts
-                  </p>
-                )}
-                <p>
-                  <b>Qualified for District Championship:</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.qualifiedDistrictCmp
-                    ? "Yes"
-                    : "No"}
-                </p>
-                <p>
-                  <b>Qualified for World Championship:</b>{" "}
-                  {advancementDetailTeam.districtRankDetails.qualifiedFirstCmp
-                    ? "Yes"
-                    : "No"}
-                </p>
-              </Container>
-            )}
-          {advancementDetailType === "regional" &&
-            advancementDetailTeam?.regionalAdvancement && (() => {
-              const r = advancementDetailTeam.regionalAdvancement;
-              return (
-                <Container fluid>
-                  <p>
-                    <b>World Rank:</b> {r.rank} ({getRegionalAdvancementTotal(r)} pts)
-                  </p>
-                  <p>
-                    <b>Championship Status:</b>{" "}
-                    {r.championshipStatus ?? "None"}
-                  </p>
-                  <p>
-                    <b>Qualified for World Championship:</b>{" "}
-                    {r.qualifiedFirstCmp ? "Yes" : "No"}
-                  </p>
-                  {r.qualifiedFirstCmpEventCode && (
-                    <p>
-                      <b>Qualified at event:</b> {r.qualifiedFirstCmpEventCode}
-                      {r.qualifiedFirstCmpDate && (
-                        <> ({moment(r.qualifiedFirstCmpDate).format("YYYY-MM-DD")})</>
-                      )}
-                    </p>
-                  )}
-                  <hr />
-                  <h6>Event breakdown</h6>
-                  {r.regional1Details && (
-                    <div className="mb-2">
-                      <b>Event 1 ({r.regional1Details.tournamentCode}):</b>{" "}
-                      {r.regional1Points ?? r.regional1Details.totalPoints} pts
-                      <ul className="small mb-0">
-                        <li>Qualification performance: {r.regional1Details.qualificationPerformancePoints}</li>
-                        <li>Alliance selection: {r.regional1Details.allianceSelectionPoints}</li>
-                        <li>Playoff advancement: {r.regional1Details.playoffAdvancementPoints}</li>
-                        <li>Award: {r.regional1Details.awardPoints}</li>
-                        <li>Team age: {r.regional1Details.teamAgePoints}</li>
-                      </ul>
-                    </div>
-                  )}
-                  {r.regional2Details && (
-                    <div className="mb-2">
-                      <b>Event 2 ({r.regional2Details.tournamentCode}):</b>{" "}
-                      {r.regional2Points ?? r.regional2Details.totalPoints} pts
-                      <ul className="small mb-0">
-                        <li>Qualification performance: {r.regional2Details.qualificationPerformancePoints}</li>
-                        <li>Alliance selection: {r.regional2Details.allianceSelectionPoints}</li>
-                        <li>Playoff advancement: {r.regional2Details.playoffAdvancementPoints}</li>
-                        <li>Award: {r.regional2Details.awardPoints}</li>
-                        <li>Team age: {r.regional2Details.teamAgePoints}</li>
-                      </ul>
-                    </div>
-                  )}
-                  {r.regionalDirectDetails && (
-                    <div className="mb-2">
-                      <b>Direct ({r.regionalDirectDetails.tournamentCode}):</b>{" "}
-                      {r.regionalDirectPoints ?? r.regionalDirectDetails.totalPoints} pts
-                    </div>
-                  )}
-                  {r.tiebreakers && (
-                    <>
-                      <hr />
-                      <h6>Tiebreakers</h6>
-                      <p className="small mb-0">
-                        Playoff: {r.tiebreakers.eventPlayoffPoints} · Alliance:{" "}
-                        {r.tiebreakers.eventAlliancePoints} · Qual:{" "}
-                        {r.tiebreakers.eventQualificationPoints}
-                        <br />
-                        Match high scores: {r.tiebreakers.matchHighScore1},{" "}
-                        {r.tiebreakers.matchHighScore2}, {r.tiebreakers.matchHighScore3}
-                      </p>
-                    </>
-                  )}
-                  {r.adjustPoints != null && r.adjustPoints !== 0 && (
-                    <p className="small mt-1">
-                      <b>Adjustment:</b> {r.adjustPoints} pts
-                    </p>
-                  )}
-                </Container>
-              );
-            })()}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleCloseAdvancementModal}>
-            Close
-          </Button>
-        </Modal.Footer>
-      </Modal>
-        );
-      })()}
+        team={advancementDetailTeam}
+        type={advancementDetailType}
+        remapNumberToString={remapNumberToString}
+      />
     </Container>
   );
 }
